+++ src/인규/브루트포스/N과M/N과M2.java
@@ -1,4 +1,4 @@
-package 인규.브루트포스;
+package 인규.브루트포스.N과M;
 
 import java.util.Scanner;
