+++ src/인규/브루트포스/N과M/N과M5.java
@@ -1,27 +1,32 @@
-package 인규.브루트포스;
+package 인규.브루트포스.N과M;
 
+import java.util.Arrays;
 import java.util.Scanner;
 
-public class N과M1 {
+public class N과M5 {
 
-  public static int[] arr;
+  public static int[] arr, nums;
   public static boolean[] visit;
+  public static int N, M;
 
   public static StringBuilder sb = new StringBuilder();
 
   public static void main(String[] args) {
     Scanner sc = new Scanner(System.in);
-    int N = sc.nextInt();
-    int M = sc.nextInt();
-
+    N = sc.nextInt();
+    M = sc.nextInt();
+    nums = new int[N];
     visit = new boolean[N];
     arr = new int[M];
-    dfs(N, M, 0);
+    for (int i = 0; i < N; i++) {
+      nums[i] = sc.nextInt();
+    }
+    Arrays.sort(nums);
+    dfs(0);
     System.out.println(sb);
-
   }
 
-  public static void dfs(int N, int M, int depth) {
+  public static void dfs(int depth) {
     if (depth == M) {
       for (int i : arr) {
         sb.append(i).append(" ");
@@ -29,12 +34,11 @@ public class N과M1 {
       sb.append("\n");
       return;
     }
-
     for (int i = 0; i < N; i++) {
       if (!visit[i]) {
         visit[i] = true;
-        arr[depth] = i + 1;
-        dfs(N, M, depth + 1);
+        arr[depth] = nums[i];
+        dfs(depth + 1);
         visit[i] = false;
       }
     }
