+++ src/인규/브루트포스/순열/차이를최대로.java
@@ -1,4 +1,4 @@
-package 인규.순열;
+package 인규.브루트포스.순열;
 
 import java.util.ArrayList;
 import java.util.Collections;
