+++ src/인규/브루트포스/재귀/암호만들기.java
@@ -1,4 +1,4 @@
-package 인규.브루트포스;
+package 인규.브루트포스.재귀;
 
 import java.util.Arrays;
 import java.util.Scanner;
